+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 * Change chart's library from `highcharts` to `chart.js`.
+* ContinueLoad with showExisting layers (odata-vector/wfs).
+* ShowAllLayerObjects doesn't clear objects on layer if showExitsting = true.
 * Style icon minus on identification container.
 * Change leafletLayer loading, moved from layer-styles to `layers-styles-editor` component.
 * Update dependency on the `jquery-minicolors` bower package to version `2.3.4`.